% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bile.R
\docType{data}
\name{bile}
\alias{bile}
\title{Bile Data}
\format{
A data frame with 29 rows and 2 variables.
}
\usage{
data(bile)
}
\description{
A dataset containing lithogenic bile concentrations in 29 Pima Indians. For 29 females, the data give age and percent saturation of bile cholesterol.
}
\details{
\itemize{
\item Age: age in years of each female indian
\item Concentration: percent saturation of bile cholesterol
}
}
\references{
Goodall, C. (1983). Examining Residuals. In Hoaglin, D. C., Mosteller, F. and Tukey, J. W. (1983). Understanding Robust and Exploratory Data Analysis. \emph{Wiley series in probability and mathematical statistics}.
}
\keyword{datasets}
