% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regL1.R
\name{class_to_regL1}
\alias{class_to_regL1}
\title{Change object class from "rq" to "regL1"}
\usage{
class_to_regL1(object)
}
\arguments{
\item{object}{Object from "rq" class.}
}
\value{
Object with class "regL1".
}
\description{
Changing the object's class from "rq" to "regL1" allows you to use functions from the 'diagL1' package normally.
}
\seealso{
\code{\link{regL1}} for fitting linear L1 models.
\code{\link{rq}} for fitting linear L1 models.
}
