% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regL1.R
\name{class_to_rq}
\alias{class_to_rq}
\title{Change object class from "regL1" to "rq"}
\usage{
class_to_rq(object)
}
\arguments{
\item{object}{Object from "regL1" class.}
}
\value{
Object with class "rq".
}
\description{
Changing the object's class from "regL1" to "rq" allows you to use functions from the 'quantreg' package normally.
}
\seealso{
\code{\link[diagL1]{regL1}} for fitting linear L1 models.
\code{\link[quantreg]{rq}} for fitting linear L1 models.
}
