% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_functions.R
\name{dlaplace}
\alias{dlaplace}
\title{Probability density function (PDF) of the Laplace distribution}
\usage{
dlaplace(x, location = 0, scale = 1)
}
\arguments{
\item{x}{Values for which to calculate the probability density.}

\item{location}{Location parameter of the Laplace distribution (default = 0).}

\item{scale}{Scale parameter of the Laplace distribution (default = 1).}
}
\value{
Vector of calculated probability densities.
}
\description{
Probability density function (PDF) of the Laplace distribution
}
\examples{
dlaplace(0, 0, 1)
}
