% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fire.R
\docType{data}
\name{Fire}
\alias{Fire}
\title{Fire Data}
\format{
A data frame with 47 rows and 6 variables.
}
\usage{
data(Fire)
}
\description{
A dataset containing incidence of fire data for 47 residential areas in Chicago for the year 1975.
}
\details{
\itemize{
\item Fire: fires per 1000 housing units
\item log_Fire: log of variable Fire
\item Theft: thefts per 1000 residents
\item Age: percent of housing units built in or before 1940
\item Income: median family income as a multiple of $1000
}
}
\references{
Birkes, D.and Dodge, Y. (1993). Alternative Methods of Regression. \emph{Jonh Wiley & Sons}.
}
\keyword{datasets}
