% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{lambda_mle}
\alias{lambda_mle}
\title{Function to estimate lambda via MLE}
\usage{
lambda_mle(model)
}
\arguments{
\item{model}{A linear model L1 fitted with the \code{rq} function from the \code{quantreg} package.}
}
\value{
lambda MLE (Maximum Likelihood Estimator) estimator.
}
\description{
Function to estimate lambda via MLE
}
\examples{
require(quantreg)
data(stackloss)
model_L1 = regL1(stack.loss ~ stack.x)
lambda_mle(model_L1)

}
