% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIF.R
\name{PIF}
\alias{PIF}
\title{Predictive Influence Function}
\usage{
PIF(y, x, w, num_cores = 2)
}
\arguments{
\item{y}{A vector with response variables.}

\item{x}{A matrix with a single explanatory variable.}

\item{w}{Explanatory variables vector for z, usually explanatory variables averages.}

\item{num_cores}{Number of cores you want to use for parallel processing (default = 2).}
}
\value{
A vector with predictive influence function value for each observation.
}
\description{
Predictive Influence Function
}
\details{
Please, install and load the "foreach" package to use this function. For more details see Rodrigues (2024).
}
\references{
Rodrigues, K. A. S. (2024). \strong{Analysis of the adjustment of the L1 regression model}.
Phd dissertation, University of São Paulo, BR.
}
