% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_functions.R
\name{plaplace}
\alias{plaplace}
\title{Cumulative distribution function (CDF) of the Laplace distribution}
\usage{
plaplace(q, location = 0, scale = 1)
}
\arguments{
\item{q}{Values for which to calculate the cumulative distribution.}

\item{location}{Location parameter of the Laplace distribution (default = 0).}

\item{scale}{Scale parameter of the Laplace distribution (default = 1).}
}
\value{
Vector of calculated cumulative distributions.
}
\description{
Cumulative distribution function (CDF) of the Laplace distribution
}
\examples{
plaplace(0, 0, 1)
}
