% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_regL1_methods.R
\name{print.forwardSearch_regL1}
\alias{print.forwardSearch_regL1}
\title{Print an forwardSearch_regL1 object}
\usage{
\method{print}{forwardSearch_regL1}(x, ...)
}
\arguments{
\item{x}{Object returned from forwardSearch_regL1.}

\item{...}{Optional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print an object generated by forwardSearch_regL1
}
\examples{
\donttest{
# applies the forward search approach to robust analysis in a linear L1 model
mod = forwardSearch_regL1(Concentration ~ Age, data = bile)
mod # or print(mod)
}

}
\seealso{
\code{\link{forwardSearch_regL1}} for apply forward search in linear L1 regression model.
}
