% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regL1.R
\name{print.regL1}
\alias{print.regL1}
\title{Print an regL1 object}
\usage{
\method{print}{regL1}(x, ...)
}
\arguments{
\item{x}{Object returned from regL1 representing the fit of the L1 model.}

\item{...}{Optional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print an object generated by regL1
}
\seealso{
\code{\link{regL1}} for fitting linear L1 models.
}
