% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_regL1_methods.R
\name{print.summary.forwardSearch_regL1}
\alias{print.summary.forwardSearch_regL1}
\title{Print Forward Search in Linear L1 Model Summary Object}
\usage{
\method{print}{summary.forwardSearch_regL1}(x, digits = 4, ...)
}
\arguments{
\item{x}{This is an object of class "\code{summary}" produced by a call to \code{summary.regL1()}.}

\item{digits}{Significant digits reported in the printed table.}

\item{...}{Optional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print summary of forward search in linear L1 model object.
}
\examples{
\donttest{
# applies the forward search approach to robust analysis in a linear L1 model
mod = forwardSearch_regL1(Concentration ~ Age, data = bile)
summary(mod)
}
}
\seealso{
\code{\link[diagL1]{forwardSearch_regL1}} for apply forward search in linear L1 regression model.
}
