% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regL1.R
\name{print.summary.regL1}
\alias{print.summary.regL1}
\title{Print Linear L1 Regression Summary Object}
\usage{
\method{print}{summary.regL1}(x, digits = max(5, .Options$digits - 2), ...)
}
\arguments{
\item{x}{This is an object of class "\code{summary.regL1}" produced by a call to \code{summary.regL1()}.}

\item{digits}{Significant digits reported in the printed table.}

\item{...}{Optional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Print summary of linear L1 regression object.
}
