% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_functions.R
\name{qlaplace}
\alias{qlaplace}
\title{Quantile function (inverse of the CDF) of the Laplace distribution}
\usage{
qlaplace(p, location = 0, scale = 1)
}
\arguments{
\item{p}{Values for which to calculate quantiles.}

\item{location}{Location parameter of the Laplace distribution (default = 0).}

\item{scale}{Scale parameter of the Laplace distribution (default = 1).}
}
\value{
Vector of calculated quantiles.
}
\description{
Quantile function (inverse of the CDF) of the Laplace distribution
}
\examples{
qlaplace(0.5, 0, 1)
}
