% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laplace_functions.R
\name{rlaplace}
\alias{rlaplace}
\title{Function to generate random values from the Laplace distribution}
\usage{
rlaplace(n, location = 0, scale = 1)
}
\arguments{
\item{n}{Number of values to generate.}

\item{location}{Location parameter of the Laplace distribution (default = 0).}

\item{scale}{Scale parameter of the Laplace distribution (default = 1).}
}
\value{
Vector of generated random values.
}
\description{
Function to generate random values from the Laplace distribution
}
\examples{
rlaplace(10, 0, 1)
}
