% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxiliary_functions.R
\name{SAE}
\alias{SAE}
\title{Function to calculate SAE}
\usage{
SAE(model)
}
\arguments{
\item{model}{A linear model L1 fitted with the \code{rq} function from the \code{quantreg} package.}
}
\value{
SAE (Sum of Absolute Errors).
}
\description{
Function to calculate SAE (Sum of Absolute Errors) of the adjusted model.
}
\examples{
require(quantreg)
data(stackloss)
model_L1 = regL1(stack.loss ~ stack.x)
SAE(model_L1)

}
