% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwd_regL1_methods.R
\name{summary.forwardSearch_regL1}
\alias{summary.forwardSearch_regL1}
\title{Summarizing Fit of Forward Search in Linear L1 Regression}
\usage{
\method{summary}{forwardSearch_regL1}(object, steps = "auto", ...)
}
\arguments{
\item{object}{Object returned from forwardSearch_regL1.}

\item{steps}{the number of forward steps to show.}

\item{...}{Optional arguments.}
}
\value{
No return value, called for side effects.
}
\description{
Returns a summary list for a forward search in linear L1 regression fit.
}
\examples{
\donttest{
# applies the forward search approach to robust analysis in a linear L1 model
mod = forwardSearch_regL1(Concentration ~ Age, data = bile)
summary(mod)
}

}
\references{
Atkinson, A.C. and Riani, M. (2000). \emph{Robust Diagnostic Regression Analysis}. New York: Springer.
}
\seealso{
\code{\link[diagL1]{forwardSearch_regL1}} for apply forward search in linear L1 regression model.
}
