% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIF.R
\name{PIF}
\alias{PIF}
\title{Predictive Influence Function}
\usage{
PIF(y, x, w, num_cores = 2)
}
\arguments{
\item{y}{A vector with response variables.}

\item{x}{A matrix with a single explanatory variable.}

\item{w}{Explanatory variables vector for z, usually explanatory variables averages.}

\item{num_cores}{Number of cores you want to use for parallel processing (default = 2).}
}
\value{
A vector with predictive influence function value for each observation.
}
\description{
Predictive Influence Function
}
\details{
Please, install and load the "foreach" package to use this function. For more details see Rodrigues (2024).
}
\references{
Rodrigues, K. A. S. (2024). \strong{Analysis of the adjustment of the L1 regression model}.
Phd dissertation, University of São Paulo, BR.

Rodrigues, K. A. S., Elian, S. N., & Pereira, G. H. A. (2025). Homoscedasticity tests for L1 regression and their performance evaluation through simulations. \emph{Statistics}, Advance online publication. https://doi.org/10.1080/02331888.2025.2536097

Rodrigues, K. A. S., & Elian, S. N. (2025). Influence measures for L1 regression: an analysis with the R package diagL1. \emph{Journal of Applied Statistics}, Advance online publication. https://doi.org/10.1080/02664763.2025.2510691
}
